+++ src/producer.rs
@@ -5,67 +5,30 @@ use crate::{
         common::Sort,
         producers::ProducersOrder,
     },
-    utils::{Images, Pagination},
+    response::Response,
+    structs::producers::Producer,
+    utils::ExternalEntry,
 };
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Producer {
-    pub mal_id: i32,
-    pub url: String,
-    pub titles: Vec<TitleEntry>,
-    pub images: Images,
-    pub favorites: i32,
-    pub count: i32,
-    pub established: Option<String>,
-    pub about: Option<String>,
-    pub external: Option<Vec<ExternalEntry>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TitleEntry {
-    #[serde(rename = "type")]
-    pub type_: String,
-    pub title: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExternalEntry {
-    pub name: String,
-    pub url: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProducerResponse<T> {
-    pub data: T,
-    pub pagination: Option<Pagination>,
-}
-
-pub enum FilterType {
-    SortFilter,
-    OrderFilter,
-    None,
-}
 
 impl JikanClient {
     pub async fn get_producer_by_id(
         &self,
         id: i32,
-    ) -> Result<ProducerResponse<Producer>, JikanError> {
+    ) -> Result<Response<Producer>, JikanError> {
         self.get(&format!("/producers/{}", id)).await
     }
 
     pub async fn get_producer_full_by_id(
         &self,
         id: i32,
-    ) -> Result<ProducerResponse<Producer>, JikanError> {
+    ) -> Result<Response<Producer>, JikanError> {
         self.get(&format!("/producers/{}/full", id)).await
     }
 
     pub async fn get_producer_external(
         &self,
         id: i32,
-    ) -> Result<ProducerResponse<Vec<ExternalEntry>>, JikanError> {
+    ) -> Result<Response<Vec<ExternalEntry>>, JikanError> {
         self.get(&format!("/producers/{}/external", id)).await
     }
 
@@ -77,7 +40,7 @@ impl JikanClient {
         order_by: Option<ProducersOrder>,
         sort: Option<Sort>,
         letter: Option<String>,
-    ) -> Result<ProducerResponse<Vec<Producer>>, JikanError> {
+    ) -> Result<Response<Vec<Producer>>, JikanError> {
         let mut params: Vec<String> = Vec::new();
 
         if let Some(p) = page {
@@ -112,7 +75,7 @@ impl JikanClient {
         self.get(&format!("/producers{}", query)).await
     }
 
-    pub async fn get_producers(&self) -> Result<ProducerResponse<Vec<Producer>>, JikanError> {
+    pub async fn get_producers(&self) -> Result<Response<Vec<Producer>>, JikanError> {
         self.get("/producers").await
     }
 }
